scr/requirements.txt
opencv-python>=4.9.0
pygrabber>=0.2
Pillow>=10.0.0